feat: add hooks for display name management, game history, lobby, and socket connection
- Introduced `useDisplayName` hook for managing and persisting user display names with validation.
- Created `useGameHistory` hook to handle game move history, time-travel viewing, and archiving completed games.
- Implemented `useLobby` hook for managing lobby matchmaking state, including join/leave operations and error handling.
- Developed `useSocketConnection` hook for managing Socket.IO connections, including lazy initialization and reconnection logic.
- Refactored `useSocketGame` to utilize new hooks for display name, game history, and lobby management, improving code organization and readability.
- Added utility functions for client ID management and room ID persistence across sessions.
