+++ server/handlers/index.js
@@ -1,0 +1,7 @@
+// Barrel file for socket handlers
+
+export { registerRoomHandlers } from './roomHandlers.js';
+export { registerGameHandlers } from './gameHandlers.js';
+export { registerVoiceHandlers } from './voiceHandlers.js';
+export { registerLobbyHandlers, handleMatch } from './lobbyHandlers.js';
+export { validateRoomId, validateDisplayName, validateIndex } from './validation.js';
