+++ src/components/BoardSquare.jsx
@@ -15,7 +15,7 @@ const BoardSquare = ({ value, onClick, isWinning, index }) => {
     if (value) return;
     setIsPressed(true);
     // Trigger haptic feedback on supported devices
-    if (navigator.vibrate) {
+    if (typeof navigator !== "undefined" && typeof navigator.vibrate === "function") {
       navigator.vibrate(10);
     }
     onClick();
@@ -26,7 +26,7 @@ const BoardSquare = ({ value, onClick, isWinning, index }) => {
     <Tooltip content={tooltipMessage}>
       <button
         type="button"
-        className={`relative flex aspect-square w-[clamp(70px,22vw,110px)] items-center justify-center rounded-2xl border border-slate-200/80 bg-white/80 text-4xl font-semibold text-slate-800 shadow-[0_15px_35px_-28px_rgba(15,23,42,0.75)] transition-all duration-200 ease-out dark:border-slate-700/70 dark:bg-slate-900/70 dark:text-slate-100 ${
+        className={`group relative flex aspect-square w-[clamp(70px,22vw,110px)] items-center justify-center rounded-2xl border border-slate-200/80 bg-white/80 text-4xl font-semibold text-slate-800 shadow-[0_15px_35px_-28px_rgba(15,23,42,0.75)] transition-all duration-200 ease-out dark:border-slate-700/70 dark:bg-slate-900/70 dark:text-slate-100 ${
           isWinning ? "ring-2 ring-indigo-400/80 dark:ring-emerald-400/70 animate-celebrate" : ""
         } ${value ? "cursor-not-allowed" : "cursor-pointer hover:-translate-y-1 hover:border-indigo-400/70 hover:bg-white hover:shadow-[0_20px_45px_-28px_rgba(99,102,241,0.6)] active:translate-y-0 active:scale-95"} ${isPressed ? "scale-95" : ""}`}
         onClick={handleClick}
