+++ src/hooks/useSocketGame.js
@@ -1,12 +1,10 @@
 import { useCallback, useEffect, useRef, useState } from "react";
 import { io } from "socket.io-client";
-import {
-  appendHistorySnapshot,
-  createSystemEntry,
-  detectChangedIndex,
-} from "../utils/history";
-import { shouldArchiveCompletedGame } from "../utils/completedGames";
-import { getDisplayName, setDisplayName as saveDisplayName } from "../utils/randomName";
+import { getDisplayName } from "../utils/randomName";
+import { getClientId, setPersistedRoom, getPersistedRoom } from "../utils/clientId";
+import useDisplayName from "./useDisplayName";
+import useLobby from "./useLobby";
+import useGameHistory from "./useGameHistory";
 
 // Local game helpers (fallback when not in a multiplayer room)
 const emptyBoard = () => Array(9).fill("");
@@ -20,6 +18,7 @@ const LINES = [
   [0, 4, 8],
   [2, 4, 6],
 ];
+
 function calcWinner(board) {
   for (const [a, b, c] of LINES) {
     if (board[a] && board[a] === board[b] && board[a] === board[c])
@@ -40,93 +39,53 @@ const initialLocalState = {
 
 export default function useSocketGame() {
   const [gameState, setGameState] = useState(initialLocalState);
-  // Move-by-move history for the CURRENT (ongoing) game
-  const [moveHistory, setMoveHistory] = useState(() => [
-    createSystemEntry({
-      board: emptyBoard(),
-      message: "Game start • X to move",
-      moveNumber: 0,
-    }),
-  ]);
-  // Concise summaries of COMPLETED games
-  const [completedGames, setCompletedGames] = useState([]); // {id, winner, draw, sequence:["X0","O4",...], totalMoves, finishedAt}
-  // Sequence of moves for current game (compact form)
-  const moveSequenceRef = useRef([]); // array of { mark, index }
-  const lastArchivedSignatureRef = useRef(null);
-  const lastBoardRef = useRef(emptyBoard());
-  // Time-travel index (which snapshot user is viewing)
-  const [viewIndex, setViewIndex] = useState(0); // 0..moveHistory.length-1
-  const viewIndexRef = useRef(0);
-  useEffect(() => {
-    viewIndexRef.current = viewIndex;
-  }, [viewIndex]);
   const [roomId, setRoomId] = useState(null);
-  const [player, setPlayer] = useState(null); // 'X' | 'O' | 'spectator'
+  const [player, setPlayer] = useState(null);
   const [roster, setRoster] = useState({ X: null, O: null, spectators: [] });
-  const [voiceRoster, setVoiceRoster] = useState({}); // { socketId: { muted: boolean } }
+  const [voiceRoster, setVoiceRoster] = useState({});
   const [isRoomCreator, setIsRoomCreator] = useState(false);
   const [message, setMessage] = useState("Local game ready");
   const [showModal, setShowModal] = useState(false);
-  const [newGameRequester, setNewGameRequester] = useState(null); // socket.id of requester
+  const [newGameRequester, setNewGameRequester] = useState(null);
   const [newGameRequestedAt, setNewGameRequestedAt] = useState(null);
-  
-  // Lobby state
-  const [lobbyQueue, setLobbyQueue] = useState([]); // [{socketId, displayName, joinedAt}]
-  const [isInLobby, setIsInLobby] = useState(false);
-  const [lobbyError, setLobbyError] = useState(null);
-  
-  // Display name state
-  const [displayName, setDisplayNameState] = useState(() => getDisplayName());
-  
-  // Connection state for UI indicator
-  const [connectionState, setConnectionState] = useState('disconnected'); // 'connected' | 'connecting' | 'disconnected'
-  
+  const [connectionState, setConnectionState] = useState("disconnected");
+
   const socketRef = useRef(null);
-  const pendingJoinRef = useRef(null); // store room code if join called before socket ready
-  // Stable clientId per browser tab/session for seat restoration across reconnects
-  const clientIdRef = useRef(null);
-  if (!clientIdRef.current) {
-    try {
-      const key = "cc_client_id";
-      let id = (typeof window !== "undefined" && window.sessionStorage)
-        ? window.sessionStorage.getItem(key)
-        : null;
-      if (!id) {
-        id = Math.random().toString(36).slice(2) + Date.now().toString(36);
-        if (typeof window !== "undefined" && window.sessionStorage) {
-          window.sessionStorage.setItem(key, id);
-        }
-      }
-      clientIdRef.current = id;
-    } catch {
-      clientIdRef.current = Math.random().toString(36).slice(2);
-    }
-  }
-  const persistedRoomKey = "cc_room_id";
-  const safeSetPersistedRoom = (value) => {
-    try {
-      if (typeof window !== "undefined" && window.sessionStorage) {
-        if (value) window.sessionStorage.setItem(persistedRoomKey, value);
-        else window.sessionStorage.removeItem(persistedRoomKey);
-      }
-    } catch {/* ignore */}
-  };
-  const safeGetPersistedRoom = () => {
-    try {
-      if (typeof window !== "undefined" && window.sessionStorage) {
-        return window.sessionStorage.getItem(persistedRoomKey);
-      }
-    } catch {/* ignore */}
-    return null;
-  };
-  // Track a one-time 'connect' listener for joinRoom to avoid stacking
+  const pendingJoinRef = useRef(null);
+  const clientIdRef = useRef(getClientId());
   const joinOnConnectHandlerRef = useRef(null);
+
+  // Use extracted hooks
+  const { displayName, updateDisplayName: baseUpdateDisplayName } = useDisplayName();
+  const {
+    moveHistory,
+    completedGames,
+    viewIndex,
+    displayedBoard,
+    recordMove,
+    finalizeCurrentGame,
+    resetHistory,
+    jumpTo,
+    resumeLatest,
+  } = useGameHistory();
+
+  const getSocket = useCallback(() => socketRef.current, []);
+  const {
+    lobbyQueue,
+    isInLobby,
+    lobbyError,
+    joinLobby,
+    leaveLobby,
+    handleLobbyUpdate,
+    handleMatchFound,
+  } = useLobby({ getSocket });
 
   const isMultiplayer = !!roomId;
 
   // Lazy socket init
   const ensureSocket = useCallback(() => {
     if (socketRef.current) return socketRef.current;
+
     const url =
       import.meta.env.VITE_SOCKET_SERVER ||
       window.location.origin.replace(/:\d+$/, ":8081");
@@ -137,57 +96,69 @@ export default function useSocketGame() {
 
     s.on("connect_error", (error) => {
       console.error("Connection error:", error);
-      setConnectionState('disconnected');
+      setConnectionState("disconnected");
     });
 
     socketRef.current = s;
 
     s.on("connect", () => {
-      setConnectionState('connected');
-      setMessage((_prev) =>
-        roomId ? `Connected as ${player || ""}` : "Connected"
-      );
-      // If there was a pending join request before socket was ready
+      setConnectionState("connected");
+      setMessage(() => (roomId ? `Connected as ${player || ""}` : "Connected"));
+
+      // Handle pending join
       if (pendingJoinRef.current) {
         const code = pendingJoinRef.current;
         pendingJoinRef.current = null;
-        s.emit("joinRoom", { roomId: code, clientId: clientIdRef.current, displayName: getDisplayName() }, (resp) => {
-          if (resp?.error) {
-            setMessage(resp.error);
-            return;
-          }
-          if (resp?.player) setPlayer(resp.player);
-        });
-      }
-      // If we were in a room before a transient disconnect, proactively re-join
-      try {
-        const saved = safeGetPersistedRoom();
-        const toJoin = roomId || saved;
-        if (toJoin) {
-          s.emit("joinRoom", { roomId: toJoin, clientId: clientIdRef.current, displayName: getDisplayName() }, (resp) => {
+        s.emit(
+          "joinRoom",
+          { roomId: code, clientId: clientIdRef.current, displayName: getDisplayName() },
+          (resp) => {
             if (resp?.error) {
               setMessage(resp.error);
-              if (resp.error === "Room not found") safeSetPersistedRoom(null);
               return;
             }
             if (resp?.player) setPlayer(resp.player);
-            if (!roomId && toJoin) setRoomId(toJoin);
-          });
+          }
+        );
+      }
+
+      // Auto-rejoin room after transient disconnect
+      try {
+        const saved = getPersistedRoom();
+        const toJoin = roomId || saved;
+        if (toJoin) {
+          s.emit(
+            "joinRoom",
+            { roomId: toJoin, clientId: clientIdRef.current, displayName: getDisplayName() },
+            (resp) => {
+              if (resp?.error) {
+                setMessage(resp.error);
+                if (resp.error === "Room not found") setPersistedRoom(null);
+                return;
+              }
+              if (resp?.player) setPlayer(resp.player);
+              if (!roomId && toJoin) setRoomId(toJoin);
+            }
+          );
         }
-      } catch {/* ignore */}
+      } catch {
+        // ignore
+      }
     });
+
     s.on("disconnect", () => {
-      setConnectionState('disconnected');
+      setConnectionState("disconnected");
       setMessage("Disconnected");
     });
 
     s.on("gameUpdate", (payload) => {
       const effectiveRoomId = payload.roomId || roomId;
-      if (effectiveRoomId) safeSetPersistedRoom(effectiveRoomId);
+      if (effectiveRoomId) setPersistedRoom(effectiveRoomId);
       setRoomId(effectiveRoomId);
       setGameState((prev) => ({ ...prev, ...payload }));
       if (payload.roster) setRoster(payload.roster);
       if (payload.voiceRoster) setVoiceRoster(payload.voiceRoster || {});
+
       if (payload.newGameRequester !== undefined) {
         setNewGameRequester(payload.newGameRequester);
         if (payload.newGameRequester) {
@@ -196,87 +167,46 @@ export default function useSocketGame() {
           setNewGameRequestedAt(null);
         }
       } else if (payload.winner) {
-        // winner state but no requester yet
         setNewGameRequester(null);
         setNewGameRequestedAt(null);
       }
+
+      // Record move in history
       const resultText = payload.winner
         ? payload.winner === "draw"
           ? "Draw"
           : `${payload.winner} wins`
         : `${payload.turn}'s turn`;
-      const boardSnapshot = Array.isArray(payload.board)
-        ? payload.board.slice()
-        : emptyBoard();
-      const changedIndex = detectChangedIndex(
-        lastBoardRef.current,
-        boardSnapshot
-      );
-      const mark =
-        changedIndex !== null && changedIndex >= 0
-          ? boardSnapshot[changedIndex]
-          : null;
+      const boardSnapshot = Array.isArray(payload.board) ? payload.board.slice() : emptyBoard();
       const entryType = payload.winner
         ? payload.winner === "draw"
           ? "draw"
           : "win"
         : "move";
-      const timestamp = Date.now();
-      setMoveHistory((historyList) => {
-        const moveNumber =
-          mark && changedIndex !== null
-            ? moveSequenceRef.current.length + 1
-            : historyList.length;
-        const { history: nextHistory, appended } = appendHistorySnapshot(
-          historyList,
-          {
-            board: boardSnapshot,
-            result: resultText,
-            type: entryType,
-            moveNumber,
-            mark,
-            index: changedIndex,
-            timestamp,
-          }
-        );
-        if (appended && mark && changedIndex !== null) {
-          moveSequenceRef.current.push({
-            mark,
-            index: changedIndex,
-            timestamp,
-          });
-        }
-        if (appended && viewIndexRef.current === historyList.length - 1) {
-          setViewIndex(nextHistory.length - 1);
-        }
-        lastBoardRef.current = boardSnapshot.slice();
-        return nextHistory;
-      });
+
+      recordMove(boardSnapshot, resultText, entryType);
       if (payload.winner) setShowModal(true);
     });
+
     s.on("gameReset", () => {
-      // Reset came from someone (maybe opponent). Clear requester flags & modal for everyone.
       setNewGameRequester(null);
       setNewGameRequestedAt(null);
       setShowModal(false);
     });
-    
-    // Lobby event listeners
-    s.on("lobbyUpdate", ({ queue }) => {
-      setLobbyQueue(queue || []);
-    });
 
+    // Lobby events
+    s.on("lobbyUpdate", handleLobbyUpdate);
     s.on("matchFound", ({ roomId: matchedRoomId, player: assignedPlayer, opponent }) => {
-      setIsInLobby(false);
+      handleMatchFound();
       setRoomId(matchedRoomId);
       setPlayer(assignedPlayer);
-      safeSetPersistedRoom(matchedRoomId);
+      setPersistedRoom(matchedRoomId);
       setMessage(`Matched with ${opponent}! You are ${assignedPlayer}`);
     });
     s.on("startGame", () => setMessage("Game started"));
 
     return s;
-  }, [player, roomId]);
+  }, [player, roomId, recordMove, handleLobbyUpdate, handleMatchFound]);
 
   const createRoom = useCallback(() => {
     const s = ensureSocket();
@@ -286,80 +216,86 @@ export default function useSocketGame() {
       setPlayer(resp.player);
       setIsRoomCreator(true);
       setMessage(`Room ${resp.roomId} created. Waiting for opponent...`);
-      safeSetPersistedRoom(resp.roomId);
+      setPersistedRoom(resp.roomId);
     });
   }, [ensureSocket]);
 
   const joinRoom = useCallback(
     (code) => {
       try {
-      const s = ensureSocket();
-      const currentDisplayName = getDisplayName();
-      const emitJoin = () => {
-        s.emit("joinRoom", { roomId: code, clientId: clientIdRef.current, displayName: currentDisplayName }, (resp) => {
-          if (resp?.error) {
-            setMessage(resp.error);
-            if (resp.error === "Room not found") safeSetPersistedRoom(null);
-            return;
-          }
-          if (resp?.player) setPlayer(resp.player);
-          setRoomId(code);
-          setIsRoomCreator(false);
-          setMessage(
-            `Joined room ${code}${
-              resp?.player === "spectator" ? " as spectator" : ""
-            }`
+        const s = ensureSocket();
+        const currentDisplayName = getDisplayName();
+        const emitJoin = () => {
+          s.emit(
+            "joinRoom",
+            { roomId: code, clientId: clientIdRef.current, displayName: currentDisplayName },
+            (resp) => {
+              if (resp?.error) {
+                setMessage(resp.error);
+                if (resp.error === "Room not found") setPersistedRoom(null);
+                return;
+              }
+              if (resp?.player) setPlayer(resp.player);
+              setRoomId(code);
+              setIsRoomCreator(false);
+              setMessage(
+                `Joined room ${code}${resp?.player === "spectator" ? " as spectator" : ""}`
+              );
+              setPersistedRoom(code);
+            }
           );
-          safeSetPersistedRoom(code);
-        });
-      };
-      if (s.connected) {
-        emitJoin();
-      } else {
-        // Ensure we emit right after the connection is established (avoid stacking listeners)
-        if (joinOnConnectHandlerRef.current) {
-          try { s.off("connect", joinOnConnectHandlerRef.current); } catch {
-            console.warn("Failed to remove previous connect listener");
-          }
-        }
-        const handler = () => {
-          emitJoin();
-          joinOnConnectHandlerRef.current = null;
         };
-        joinOnConnectHandlerRef.current = handler;
-        s.once("connect", handler);
-        // Kick off connection if not already
-        try { s.connect(); } catch {
-          console.warn("Socket connect() failed");
-        }
-        // If it connected in-between, emit now
+
         if (s.connected) {
           emitJoin();
-          try { s.off("connect", handler); } catch {
-            console.warn("Failed to remove connect listener");
+        } else {
+          if (joinOnConnectHandlerRef.current) {
+            try {
+              s.off("connect", joinOnConnectHandlerRef.current);
+            } catch {
+              console.warn("Failed to remove previous connect listener");
+            }
           }
-          joinOnConnectHandlerRef.current = null;
+          const handler = () => {
+            emitJoin();
+            joinOnConnectHandlerRef.current = null;
+          };
+          joinOnConnectHandlerRef.current = handler;
+          s.once("connect", handler);
+          try {
+            s.connect();
+          } catch {
+            console.warn("Socket connect() failed");
+          }
+          if (s.connected) {
+            emitJoin();
+            try {
+              s.off("connect", handler);
+            } catch {
+              console.warn("Failed to remove connect listener");
+            }
+            joinOnConnectHandlerRef.current = null;
+          }
         }
+      } catch (e) {
+        console.error("Join room error:", e);
       }
-      } catch (e) { console.error("Join room error:", e); };
     },
     [ensureSocket]
   );
 
   const handleSquareClick = useCallback(
     (index) => {
-      // If user is viewing a past move, auto-resume latest before applying a move
-      setViewIndex((currentIdx) => {
-        const latest = moveHistory.length - 1;
-        return currentIdx === latest ? currentIdx : latest;
-      });
+      // Auto-resume to latest if viewing past
+      resumeLatest();
+
       // Multiplayer path
       if (isMultiplayer) {
         if (!socketRef.current) return;
-        if (player === "spectator") return; // read-only
-        // Prevent move if it's not this player's turn
+        if (player === "spectator") return;
         if (gameState.turn !== player) return;
-        // Optimistic update for snappy UX; server will reconcile via gameUpdate
+
+        // Optimistic update
         setGameState((curr) => {
           if (curr.winner || curr.board[index] !== "") return curr;
           const board = curr.board.slice();
@@ -371,49 +307,10 @@ export default function useSocketGame() {
               ? "Draw"
               : `${result.winner} wins`
             : `${nextTurn}'s turn`;
-          const changedIndex = detectChangedIndex(
-            lastBoardRef.current,
-            board
-          );
-          const mark =
-            changedIndex !== null && changedIndex >= 0 ? board[changedIndex] : null;
-          const entryType = result
-            ? result.winner === "draw"
-              ? "draw"
-              : "win"
-            : "move";
-          const timestamp = Date.now();
-          // History optimistic append with de-duplication
-          setMoveHistory((historyList) => {
-            const moveNumber =
-              mark && changedIndex !== null
-                ? moveSequenceRef.current.length + 1
-                : historyList.length;
-            const { history: nextHistory, appended } = appendHistorySnapshot(
-              historyList,
-              {
-                board,
-                result: resultText,
-                type: entryType,
-                moveNumber,
-                mark,
-                index: changedIndex,
-                timestamp,
-              }
-            );
-            if (appended && mark && changedIndex !== null) {
-              moveSequenceRef.current.push({
-                mark,
-                index: changedIndex,
-                timestamp,
-              });
-            }
-            if (appended && viewIndexRef.current === historyList.length - 1) {
-              setViewIndex(nextHistory.length - 1);
-            }
-            lastBoardRef.current = board.slice();
-            return nextHistory;
-          });
+          const entryType = result ? (result.winner === "draw" ? "draw" : "win") : "move";
+
+          recordMove(board, resultText, entryType);
+
           return {
             ...curr,
             board,
@@ -422,9 +319,11 @@ export default function useSocketGame() {
             winningLine: result ? result.line : curr.winningLine,
           };
         });
+
         socketRef.current.emit("makeMove", { roomId, index });
         return;
       }
+
       // Local fallback
       setGameState((current) => {
         if (current.winner || current.board[index] !== "") return current;
@@ -432,6 +331,7 @@ export default function useSocketGame() {
         board[index] = current.turn;
         const result = calcWinner(board);
         let next = { ...current, board };
+
         if (result) {
           next.winner = result.winner;
           next.winningLine = result.line;
@@ -441,97 +341,26 @@ export default function useSocketGame() {
         } else {
           next.turn = current.turn === "X" ? "O" : "X";
         }
+
         const resultText = result
           ? result.winner === "draw"
             ? "Draw"
             : `${result.winner} wins`
           : `${next.turn}'s turn`;
-        const changedIndex = detectChangedIndex(lastBoardRef.current, board);
-        const mark =
-          changedIndex !== null && changedIndex >= 0 ? board[changedIndex] : null;
-        const entryType = result
-          ? result.winner === "draw"
-            ? "draw"
-            : "win"
-          : "move";
-        const timestamp = Date.now();
-        setMoveHistory((historyList) => {
-          const moveNumber =
-            mark && changedIndex !== null
-              ? moveSequenceRef.current.length + 1
-              : historyList.length;
-          const { history: nextHistory, appended } = appendHistorySnapshot(
-            historyList,
-            {
-              board,
-              result: resultText,
-              type: entryType,
-              moveNumber,
-              mark,
-              index: changedIndex,
-              timestamp,
-            }
-          );
-          if (appended && mark && changedIndex !== null) {
-            moveSequenceRef.current.push({
-              mark,
-              index: changedIndex,
-              timestamp,
-            });
-          }
-          if (appended && viewIndexRef.current === historyList.length - 1)
-            setViewIndex(nextHistory.length - 1);
-          lastBoardRef.current = board.slice();
-          return nextHistory;
-        });
+        const entryType = result ? (result.winner === "draw" ? "draw" : "win") : "move";
+
+        recordMove(board, resultText, entryType);
         return next;
       });
     },
-    // include gameState.turn to satisfy exhaustive-deps, but we only read it synchronously
-    [isMultiplayer, player, roomId, moveHistory.length, gameState.turn]
+    [isMultiplayer, player, roomId, gameState.turn, recordMove, resumeLatest]
   );
 
-  const finalizeCurrentGameIfFinished = useCallback(() => {
-    if (!gameState.winner) return; // only store completed
-
-    const sequenceStrings = moveSequenceRef.current.map(
-      (m) => `${m.mark ?? ""}${m.index ?? ""}`
-    );
-    const lastMoveTimestamp =
-      moveSequenceRef.current.length > 0
-        ? moveSequenceRef.current[moveSequenceRef.current.length - 1].timestamp
-        : null;
-
-    const { shouldArchive, signature } = shouldArchiveCompletedGame({
-      lastSignature: lastArchivedSignatureRef.current,
-      sequenceStrings,
-      lastMoveTimestamp,
-    });
-
-    if (!shouldArchive) {
-      lastArchivedSignatureRef.current = signature;
-      return;
-    }
-
-    const summary = {
-      id: Date.now(),
-      winner: gameState.winner === "draw" ? null : gameState.winner,
-      draw: gameState.winner === "draw",
-      sequence: sequenceStrings,
-      totalMoves: sequenceStrings.length,
-      finishedAt: new Date().toISOString(),
-    };
-
-    setCompletedGames((g) => [...g, summary]);
-    lastArchivedSignatureRef.current = signature;
-  }, [gameState.winner]);
-
   const resetGame = useCallback(() => {
-    // If current game finished, persist summary
-    finalizeCurrentGameIfFinished();
+    finalizeCurrentGame(gameState.winner);
+
     if (isMultiplayer && socketRef.current) {
       socketRef.current.emit("resetGame", { roomId });
-      // After server reset, local state will update via event; prime local trackers
     } else {
       setGameState((s) => ({
         ...initialLocalState,
@@ -539,87 +368,53 @@ export default function useSocketGame() {
         oScore: s.oScore,
       }));
     }
-    // Reset per-game tracking
-    const freshBoard = emptyBoard();
-    moveSequenceRef.current = [];
-    lastBoardRef.current = freshBoard;
-    setMoveHistory([
-      createSystemEntry({
-        board: freshBoard,
-        message: "New game • X to move",
-        moveNumber: 0,
-        tag: "reset",
-      }),
-    ]);
-    setViewIndex(0);
+
+    resetHistory("New game • X to move", "reset");
     setShowModal(false);
     setNewGameRequester(null);
-  }, [finalizeCurrentGameIfFinished, isMultiplayer, roomId]);
+  }, [finalizeCurrentGame, gameState.winner, isMultiplayer, roomId, resetHistory]);
 
   const resetScores = useCallback(() => {
     if (isMultiplayer && socketRef.current) {
       socketRef.current.emit("resetScores", { roomId });
       return;
     }
-    setGameState((_s) => ({ ...initialLocalState }));
-    const freshBoard = emptyBoard();
-    moveSequenceRef.current = [];
-    lastBoardRef.current = freshBoard;
-    setMoveHistory([
-      createSystemEntry({
-        board: freshBoard,
-        message: "Scores reset • X to move",
-        moveNumber: 0,
-        tag: "system",
-      }),
-    ]);
-    setViewIndex(0);
+    setGameState(() => ({ ...initialLocalState }));
+    resetHistory("Scores reset • X to move", "system");
     setShowModal(false);
     setNewGameRequester(null);
-  }, [isMultiplayer, roomId]);
+  }, [isMultiplayer, roomId, resetHistory]);
 
   const leaveRoom = useCallback(() => {
-    // Return a promise so callers can await before navigating
     return new Promise((resolve) => {
-      // Snapshot current room before clearing for server ack
       const currentRoomId = roomId;
 
       const doLocalCleanup = () => {
-        try { finalizeCurrentGameIfFinished(); } catch { /* ignore */ }
+        try {
+          finalizeCurrentGame(gameState.winner);
+        } catch {
+          // ignore
+        }
         setRoomId(null);
         setPlayer(null);
         setIsRoomCreator(false);
         setMessage("Left room");
         setGameState(initialLocalState);
         setRoster({ X: null, O: null, spectators: [] });
-        moveSequenceRef.current = [];
-        const freshBoard = emptyBoard();
-        lastBoardRef.current = freshBoard;
-        setMoveHistory([
-          createSystemEntry({
-            board: freshBoard,
-            message: "Left room",
-            moveNumber: 0,
-            tag: "system",
-          }),
-        ]);
-        setViewIndex(0);
+        resetHistory("Left room", "system");
         setShowModal(false);
         setNewGameRequester(null);
-        safeSetPersistedRoom(null);
+        setPersistedRoom(null);
       };
 
-      // If not in multiplayer or socket missing, just cleanup and resolve
       if (!isMultiplayer || !socketRef.current || !currentRoomId) {
         doLocalCleanup();
         resolve();
         return;
       }
 
-      // Optimistically cleanup first to avoid URL flicker from auto-redirect effect
       doLocalCleanup();
 
-      // Inform server; resolve regardless of ack result to not block UI
       try {
         socketRef.current.emit(
           "leaveRoom",
@@ -636,48 +431,7 @@ export default function useSocketGame() {
         resolve();
       }
     });
-  }, [isMultiplayer, roomId, finalizeCurrentGameIfFinished]);
-
-  // Join matchmaking lobby
-  const joinLobby = useCallback((displayName) => {
-    return new Promise((resolve, reject) => {
-      const s = ensureSocket();
-      if (!s) {
-        reject(new Error('Socket not available'));
-        return;
-      }
-
-      setLobbyError(null);
-      
-      s.emit('joinLobby', { displayName }, (response) => {
-        if (response?.success) {
-          setIsInLobby(true);
-          resolve(response);
-        } else {
-          const error = response?.error || 'Failed to join lobby';
-          setLobbyError(error);
-          reject(new Error(error));
-        }
-      });
-    });
-  }, [ensureSocket]);
-
-  // Leave matchmaking lobby
-  const leaveLobby = useCallback(() => {
-    return new Promise((resolve) => {
-      const s = socketRef.current;
-      if (!s) {
-        resolve();
-        return;
-      }
-
-      s.emit('leaveLobby', (response) => {
-        setIsInLobby(false);
-        setLobbyError(null);
-        resolve(response);
-      });
-    });
-  }, []);
+  }, [isMultiplayer, roomId, finalizeCurrentGame, gameState.winner, resetHistory]);
 
   // Cleanup socket on unmount
   useEffect(
@@ -687,13 +441,26 @@ export default function useSocketGame() {
     []
   );
 
+  // Display name update with server notification
+  const updateDisplayName = useCallback(
+    (newName) => {
+      return baseUpdateDisplayName(newName, {
+        onServerNotify: (trimmed) => {
+          if (socketRef.current && roomId) {
+            socketRef.current.emit("updateDisplayName", { roomId, displayName: trimmed });
+          }
+        },
+      });
+    },
+    [baseUpdateDisplayName, roomId]
+  );
+
   return {
     gameState,
     history: moveHistory,
     completedGames,
     viewIndex,
-    displayedBoard:
-      moveHistory[Math.min(viewIndex, moveHistory.length - 1)].squares,
+    displayedBoard,
     message,
     roomId,
     player,
@@ -706,7 +473,6 @@ export default function useSocketGame() {
     newGameRequester,
     requestNewGame: () => {
       if (!isMultiplayer || !socketRef.current) return;
-      // Set requester locally for instant UI feedback
       setNewGameRequester(socketRef.current.id);
       setNewGameRequestedAt(Date.now());
       socketRef.current.emit("requestNewGame", { roomId });
@@ -726,11 +492,8 @@ export default function useSocketGame() {
     resetGame,
     resetScores,
     leaveRoom,
-    jumpTo: (idx) => {
-      if (idx < 0 || idx >= moveHistory.length) return;
-      setViewIndex(idx);
-    },
-    resumeLatest: () => setViewIndex(moveHistory.length - 1),
+    jumpTo,
+    resumeLatest,
     // Connection state
     connectionState,
     // Lobby methods
@@ -741,16 +504,6 @@ export default function useSocketGame() {
     leaveLobby,
     // Display name
     displayName,
-    updateDisplayName: (newName) => {
-      const trimmed = String(newName || '').trim();
-      if (trimmed.length < 2 || trimmed.length > 20) return false;
-      saveDisplayName(trimmed);
-      setDisplayNameState(trimmed);
-      // Notify server if in a room
-      if (socketRef.current && roomId) {
-        socketRef.current.emit("updateDisplayName", { roomId, displayName: trimmed });
-      }
-      return true;
-    },
+    updateDisplayName,
   };
 }
